sql parser improved support db2 merge syntax. for issue #1324
