+++ src/main/java/com/alibaba/druid/sql/dialect/db2/parser/DB2Lexer.java
@@ -40,6 +40,9 @@ public class DB2Lexer extends Lexer {
         map.put("CONCAT", Token.CONCAT);
         map.put("CONTINUE", Token.CONTINUE);
         map.put("IDENTITY", Token.IDENTITY);
+        map.put("MERGE", Token.MERGE);
+        map.put("USING", Token.USING);
+        map.put("MATCHED", Token.MATCHED);
         
         DEFAULT_DB2_KEYWORDS = new Keywords(map);
     }
