+++ src/test/java/com/alibaba/druid/bvt/sql/db2/DB2MergeTest.java
@@ -1,0 +1,76 @@
+/*
+ * Copyright 1999-2101 Alibaba Group Holding Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.alibaba.druid.bvt.sql.db2;
+
+import java.util.List;
+
+import org.junit.Assert;
+
+import com.alibaba.druid.sql.DB2Test;
+import com.alibaba.druid.sql.SQLUtils;
+import com.alibaba.druid.sql.ast.SQLStatement;
+import com.alibaba.druid.sql.dialect.db2.parser.DB2StatementParser;
+import com.alibaba.druid.sql.dialect.db2.visitor.DB2SchemaStatVisitor;
+import com.alibaba.druid.stat.TableStat;
+import com.alibaba.druid.stat.TableStat.Column;
+import com.alibaba.druid.util.JdbcConstants;
+
+public class DB2MergeTest extends DB2Test {
+
+    public void test_0() throws Exception {
+        String sql = "MERGE INTO product AS T "
+                + "      USING sales AS S"
+                + "    ON S.id = T.id"
+                + "  WHEN MATCHED THEN UPDATE SET inventory = T.inventory - S.sold;"
+                ;
+
+        DB2StatementParser parser = new DB2StatementParser(sql);
+        List<SQLStatement> statementList = parser.parseStatementList();
+        SQLStatement stmt = statementList.get(0);
+        print(statementList);
+
+        Assert.assertEquals(1, statementList.size());
+
+        DB2SchemaStatVisitor visitor = new DB2SchemaStatVisitor();
+        stmt.accept(visitor);
+
+//        System.out.println("Tables : " + visitor.getTables());
+//        System.out.println("fields : " + visitor.getColumns());
+//        System.out.println("coditions : " + visitor.getConditions());
+//        System.out.println("orderBy : " + visitor.getOrderByColumns());
+
+        Assert.assertEquals(2, visitor.getTables().size());
+        Assert.assertEquals(4, visitor.getColumns().size());
+        Assert.assertEquals(2, visitor.getConditions().size());
+
+        Assert.assertTrue(visitor.getTables().containsKey(new TableStat.Name("sales")));
+        Assert.assertTrue(visitor.getTables().containsKey(new TableStat.Name("product")));
+
+//         Assert.assertTrue(visitor.getColumns().contains(new Column("DSN8B10.EMP", "WORKDEPT")));
+        // Assert.assertTrue(visitor.getColumns().contains(new Column("mytable", "first_name")));
+        // Assert.assertTrue(visitor.getColumns().contains(new Column("mytable", "full_name")));
+
+        Assert.assertEquals("MERGE INTO product T"
+                + "\nUSING sales S ON (S.id = T.id) "
+                + "\nWHEN MATCHED THEN UPDATE SET inventory = T.inventory - S.sold", //
+                            SQLUtils.toSQLString(stmt, JdbcConstants.DB2));
+        
+        Assert.assertEquals("merge into product T"
+                + "\nusing sales S on (S.id = T.id) "
+                + "\nwhen matched then update set inventory = T.inventory - S.sold", //
+                            SQLUtils.toSQLString(stmt, JdbcConstants.DB2, SQLUtils.DEFAULT_LCASE_FORMAT_OPTION));
+    }
+}
