+++ src/test/java/com/alibaba/druid/bvt/sql/oracle/OracleMergeTest8.java
@@ -22,7 +22,7 @@ import org.junit.Assert;
 import com.alibaba.druid.sql.OracleTest;
 import com.alibaba.druid.sql.SQLUtils;
 import com.alibaba.druid.sql.ast.SQLStatement;
-import com.alibaba.druid.sql.dialect.oracle.ast.stmt.OracleMergeStatement;
+import com.alibaba.druid.sql.ast.statement.SQLMergeStatement;
 import com.alibaba.druid.sql.dialect.oracle.parser.OracleStatementParser;
 import com.alibaba.druid.sql.parser.SQLStatementParser;
 
@@ -40,7 +40,7 @@ public class OracleMergeTest8 extends OracleTest {
 
         SQLStatementParser parser = new OracleStatementParser(sql);
         List<SQLStatement> stmtList = parser.parseStatementList();
-        OracleMergeStatement mergeStatement = (OracleMergeStatement) stmtList.get(0);
+        SQLMergeStatement mergeStatement = (SQLMergeStatement) stmtList.get(0);
         String result = SQLUtils.toOracleString(mergeStatement);
         Assert.assertEquals("MERGE INTO (SELECT *"
                 + "\n\tFROM T09_RULE_CAL_COUNT"
